Create game-specific Scene2.5D for drawing pips
