+++ src/OpenSage.Mods.Bfme/BfmeDefinition.cs
@@ -44,6 +44,8 @@ namespace OpenSage.Mods.Bfme
             return new OnDemandAssetLoadStrategy(PathResolvers.W3d, PathResolvers.BfmeTexture);
         }
 
+        public Scene25D CreateScene25D(Scene3D scene3D, AssetStore assetStore) => new(scene3D, assetStore);
+
         public static BfmeDefinition Instance { get; } = new BfmeDefinition();
 
         public string LauncherExecutable => "lotrbfme.exe";
